componentFitting function for equivalent circuits
A function for simulating serial components of an equivalent circuit separately. Useful for visualizing the behaviour of different parts of an equivalent circuit.

Ref enhancement: Visualization of fit components

The function takes a string and associated input parameters, and simulates the complete circuit as well as the serial parts of the parent circuit.
The function currently uses CustomCircuit.predict() to simulate the circuits.